backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
websockets==12.0
python-dotenv==1.0.0
anthropic==0.18.0
httpx==0.27.2